chore: move browser-pool files up to lib root directory
