+++ lib/browser-pool/basic-pool.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const _ = require('lodash');
-const Browser = require('../../browser/new-browser');
-const CancelledError = require('../errors/cancelled-error');
-const Events = require('../../constants/runner-events');
+const Browser = require('../browser/new-browser');
+const CancelledError = require('../core/errors/cancelled-error');
+const Events = require('../constants/runner-events');
 const Pool = require('./pool');
 const debug = require('debug');
