+++ lib/browser-pool/caching-pool.js
