+++ lib/browser-pool/limited-pool.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const yallist = require('yallist');
 const Pool = require('./pool');
-const CancelledError = require('../errors/cancelled-error');
+const CancelledError = require('../core/errors/cancelled-error');
 const debug = require('debug');
 const {buildCompositeBrowserId} = require('./utils');
