+++ lib/browser-pool/limited-use-set.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const _ = require('lodash');
-const Promise = require('bluebird');
 const debug = require('debug');
 
 /**
