+++ lib/browser-pool/pool.js
