+++ lib/browser-pool/utils.js
