+++ test/lib/browser-pool/utils.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const utils = require('lib/core/browser-pool/utils');
+const utils = require('lib/browser-pool/utils');
 
 describe('browser-pool/utils', () => {
     describe('buildCompositeBrowserId', () => {
